README
In the realm of Software as a Service (SaaS), a company specializing in sales and marketing software for B2B clients has amassed a rich dataset stored in SaaS-Sales.csv. They've enlisted a data scientist to unlock hidden insights within the data. This project aims to optimize marketing, enhance sales performance, and strengthen their position as an industry leader. The outcomes will not only benefit the company but also provide valuable insights to their clients, fostering a mutually beneficial partnership. With data as their compass, they're navigating the competitive landscape with innovation and confidence.